swagger2openapi: enhancements to form/body parameter conversions
Fixes #112 #113

Why:
When converting swagger 2.0 apispecs to openapi 3,
swagger2openapi is not currently copying extensions found
on form and body parameters.  Extensions found on form parameters
should (IMO) be copied to the resulting schema property in the
converted openapi3 apispec.
Likewise, extensions found on body parameters should (IMO) be copied
to the resulting requestBody within the converted apispec.
Along with this issue re: extensions, it would be nice if swagger2openapi
could optionally set an extension to preserve the body parameter name
from the input swagger 2.0 apispec.

What:
This PR adds support for copying extensions from form and body parameters
to the appropriate location in the converted apispec.   Also, it
adds support for a new "--rbname" command line option which the user can
specify to indicate the name of the extension to use within the converted
apispec to preserve the body parameter name from the input apispec.
This extension will be set on the operation within the converted apispec.
I suspect a common use of this will be that folks who are using code
generators based on openapi-generator will use:
    --rbname x-codegen-request-body-name